lightout/requirements.txt
Babel
dash
dash-bootstrap-components
pandas
pandas-stubs
plotly
pyyaml
fastf1
python-i18n[YAML]
sqlalchemy
pymongo
gunicorn